Revert "initial changes to rename realm"
This reverts commit d18b411448bec5d31ade91fc4c7edb3c9e06cd52.
