+++ Makefile
@@ -7,10 +7,6 @@ COVERAGE=coverage.out
 export PATH := ./bin:$(PATH)
 export GO111MODULE := on
 
-.PHONY: build
-build:
-	go install $(SOURCE_FILES)
-	
 .PHONY: setup
 setup:  ## Install dev tools
 	@echo "==> Installing dependencies..."
