+++ realm/realm.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package appservices // import "go.mongodb.org/realm/realm"
+package realm // import "go.mongodb.org/realm/realm"
 
 import (
 	"bytes"
@@ -28,7 +28,6 @@ import (
 	"strings"
 
 	"github.com/google/go-querystring/query"
-
 	"go.mongodb.org/atlas/mongodbatlas"
 )
